moved the logic to one central location (shared/logutil/logutil.go), removed new line chars from file logs
