+++ shared/logutil/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["logutil.go"],
+    importpath = "github.com/prysmaticlabs/prysm/shared/logutil",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_joonix_log//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+        "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
+    ],
+)
