+++ validator/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
         "//shared/featureconfig:go_default_library",
+        "//shared/logutil:go_default_library",
         "//shared/version:go_default_library",
         "//validator/accounts:go_default_library",
         "//validator/node:go_default_library",
